Apply automatic Clippy fixes

+++ src/compiler.rs
@@ -80,17 +80,14 @@ impl<'a> Compiler<'a> {
 
         stderr.rewind().unwrap();
 
-        match result {
-            Some(status) => {
-                if status.code().expect("The compiler returned an invalid status code") != 0 {
-                    let compilation_result = read_to_string(stderr).expect("Failed to read compiler output");
-                    return Err(compilation_result);
-                }
+        if let Some(status) = result {
+            if status.code().expect("The compiler returned an invalid status code") != 0 {
+                let compilation_result = read_to_string(stderr).expect("Failed to read compiler output");
+                return Err(compilation_result);
             }
-            None => {
-                child.kill().unwrap();
-                return Err("Compilation timed out".to_string());
-            }
+        } else {
+            child.kill().unwrap();
+            return Err("Compilation timed out".to_string());
         }
         Ok(time_before_compilation.elapsed())
     }
@@ -109,7 +106,7 @@ impl<'a> Compiler<'a> {
         name: &'static str,
     ) -> Result<(PathBuf, Option<Duration>), CompilerError> {
         debug_assert!(PathBuf::from(name).extension().is_none());
-        let output_path = self.tempdir.path().join(format!("{}.o", name));
+        let output_path = self.tempdir.path().join(format!("{name}.o"));
 
         if !Self::is_source_file(source_path) {
             fs::copy(source_path, &output_path).expect("The provided filename is invalid");
